changing the telegram_id type to BigInteger

+++ app/models.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, String, Integer, ARRAY
+from sqlalchemy import Column, String, Integer, ARRAY, BigInteger
 
 from app.database import Base
 
 
 class User(Base):
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     username = Column(String, nullable=True)
     location  = Column(String, nullable=False)
     language = Column(String, nullable=False)
